% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_rstatsjobsbot.R
\name{get_rtable_posts}
\alias{get_rtable_posts}
\title{Get retweetable tweets.}
\usage{
get_rtable_posts(user, from_time)
}
\arguments{
\item{user}{A character with the bot username.}

\item{from_time}{A POSIXct indicating minimum tweets time to attend.}
}
\description{
Returns a table of tweets to retweet.
}
