% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rstatsjobsbot.R
\name{random_job_message}
\alias{random_job_message}
\title{Returns random messages.}
\usage{
random_job_message(n)
}
\arguments{
\item{n}{A numeric indicating the number of messages to return.}
}
\description{
Return a message to add to retweet with comments.
}
