% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_rstatsjobsbot.R
\name{search_tweets_with_user_data}
\alias{search_tweets_with_user_data}
\title{Search Tweets With User Data.}
\usage{
search_tweets_with_user_data(q, type = "recent", include_rts = FALSE)
}
\arguments{
\item{q}{Query to be searched, see `?rtweet::search_tweets`.}

\item{type}{Which type of search results to return, see `?rtweet::search_tweets`.}

\item{include_rts}{Logical, indicating whether to include retweets in search results, see
`?rtweet::search_tweets`.}
}
\description{
Search tweets, and return them with users data.
}
