% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_rstatsjobsbot.R
\name{start_rstatsjobsbot}
\alias{start_rstatsjobsbot}
\title{Start RStatsJobsBot bot server.}
\usage{
start_rstatsjobsbot(
  user = "@RStatsJobsBot",
  from_time = Sys.time(),
  sleep_secs = 15 * 60,
  rtweet_token_file = NULL
)
}
\arguments{
\item{user}{A character with the bot username.}

\item{from_time}{A POSIXct indicating minimum tweets time to attend.}

\item{sleep_secs}{Numeric indicating the seconds to sleep between execution.}

\item{rtweet_token_file}{A character with the file path of an rds file with
the rtweet token to be used. If NULL, it will load the default token.}
}
\description{
Starts the server, will be executing until it is aborted Ctrl+c .
}
