% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_rstatsjobsbot.R
\name{start_rstatsjobsbot}
\alias{start_rstatsjobsbot}
\title{Start RStatsJobsBot bot server.}
\usage{
start_rstatsjobsbot(
  rtweet_app,
  rtweet_consumer_key,
  rtweet_consumer_secret,
  rtweet_access_token,
  rtweet_access_secret,
  user = "RStatsJobsBot",
  from_time = Sys.time(),
  max_hashtags = 15
)
}
\arguments{
\item{rtweet_app}{Twitter credentials.}

\item{rtweet_consumer_key}{Twitter credentials.}

\item{rtweet_consumer_secret}{Twitter credentials.}

\item{rtweet_access_token}{Twitter credentials.}

\item{rtweet_access_secret}{Twitter credentials.}

\item{user}{A character with the bot username.}

\item{from_time}{A POSIXct indicating minimum tweets time to attend.}

\item{max_hashtags}{Numeric indicating the max number of hashtags that a
tweet can have (removed if has more).}
}
\description{
Starts the server, will be executing until it is aborted Ctrl+c .
}
